refactor(website): inject the icon from name-gen level
Remove the hard-coding from summoning circle component
